minor esthetic fixes

+++ include/statistics.hpp
@@ -25,19 +25,19 @@ struct statistics {
         void init(size_t vocab_size) {
             assert(vocab_size);
             m_vocab_size = vocab_size;
-            uint8_t N = m_config.max_order;
-            for (uint8_t n = 1; n < N; ++n) {
+            uint64_t N = m_config.max_order;
+            for (uint64_t n = 1; n < N; ++n) {
                 m_tmp_stats.resize(n, m_vocab_size);
             }
             m_tmp_data.probs_offsets.resize(N, std::vector<uint64_t>(0, 0));
-            for (uint8_t n = 2; n <= N; ++n) {
+            for (uint64_t n = 2; n <= N; ++n) {
                 m_tmp_data.probs_offsets[n - 1].resize(m_vocab_size, 0);
             }
         }
 
         template <typename Iterator>
         void compute_left_extensions(Iterator begin, size_t len) {
-            uint8_t N = m_config.max_order;
+            uint64_t N = m_config.max_order;
             m_num_ngrams[N - 1] += len;
 
             auto prev_ptr = *begin;
@@ -49,7 +49,7 @@ struct statistics {
                 auto ptr = *begin;
                 word_id right = ptr[N - 1];
 
-                for (uint8_t n = 1; n < N; ++n) {
+                for (uint64_t n = 1; n < N; ++n) {
                     bool context_changes =
                         !ptr.equal_to(prev_ptr, N - n, N - 1);
                     if (n != 1 and context_changes) {
@@ -79,12 +79,12 @@ struct statistics {
 
         void finalize() {
             ++m_num_ngrams[3];
-            for (uint8_t n = 2; n < m_config.max_order; ++n) {
+            for (uint64_t n = 2; n < m_config.max_order; ++n) {
                 ++m_num_ngrams[n - 2];
                 m_tmp_stats.combine(n);
                 m_tmp_stats.release(n);
             }
-            for (uint8_t n = 2; n <= m_config.max_order; ++n) {
+            for (uint64_t n = 2; n <= m_config.max_order; ++n) {
                 for (uint64_t k = 1; k <= 4; ++k) {
                     m_t[n - 1][k - 1] = m_tmp_stats.t[n - 1][k - 1];
                 }
@@ -92,15 +92,15 @@ struct statistics {
         }
 
         void build(statistics& stats) {
-            uint8_t N = m_config.max_order;
+            uint64_t N = m_config.max_order;
             stats.num_ngrams(1) = m_vocab_size;
-            for (uint8_t n = 2; n <= N; ++n) {
+            for (uint64_t n = 2; n <= N; ++n) {
                 stats.num_ngrams(n) = m_num_ngrams[n - 1];
             }
 
             std::cerr << "number of ngrams:\n";
             size_t sum = 0;
-            for (uint8_t n = 1; n <= N; ++n) {
+            for (uint64_t n = 1; n <= N; ++n) {
                 std::cerr << int(n) << "-grams: " << stats.num_ngrams(n)
                           << "\n";
                 sum += stats.num_ngrams(n);
@@ -113,7 +113,7 @@ struct statistics {
                 if (k <= 4) ++m_t[0][k - 1];
             }
 
-            for (uint8_t n = 2; n <= N; ++n) {
+            for (uint64_t n = 2; n <= N; ++n) {
                 auto& positions = m_tmp_data.probs_offsets[n - 1];
                 // compute prefix sums
                 for (uint64_t id = 0, sum = 0; id < m_vocab_size; ++id) {
@@ -128,7 +128,7 @@ struct statistics {
             }
 
             // NOTE: do not compute for small synthetic datasets
-            for (uint8_t n = 1; n <= N; ++n) {
+            for (uint64_t n = 1; n <= N; ++n) {
                 for (uint64_t k = 1; k <= 4; ++k) {
                     try {
                         D(n, k) = compute_discount(n, k);
@@ -169,20 +169,20 @@ struct statistics {
         float m_unk_prob;  // prob of <unk> word, which is backoff(empty) /
                            // vocabulary_size
 
-        float& D(uint8_t n, uint64_t k) {
+        float& D(uint64_t n, uint64_t k) {
             assert(k > 0);
             assert(n >= 1 and n <= m_config.max_order);
             if (k >= 3) return m_D[n - 1].back();
             return m_D[n - 1][k - 1];
         }
 
-        inline uint64_t t(uint8_t n, uint64_t k) const {
+        inline uint64_t t(uint64_t n, uint64_t k) const {
             assert(n >= 1 and n <= m_config.max_order);
             assert(k > 0 and k <= 4);
             return m_t[n - 1][k - 1];
         }
 
-        float compute_discount(uint8_t n, uint64_t k) {
+        float compute_discount(uint64_t n, uint64_t k) {
             assert(k > 0 and k <= 4);
             assert(n >= 1 and n <= m_config.max_order);
             if (k <= 3) {
@@ -194,8 +194,8 @@ struct statistics {
             return compute_discount(n, 3);
         }
 
-        void complain(uint8_t n, uint64_t k) {
-            auto check = [&](uint8_t n, uint64_t k) {
+        void complain(uint64_t n, uint64_t k) {
+            auto check = [&](uint64_t n, uint64_t k) {
                 if (!t(n, k)) std::cerr << k << "\n";
             };
             std::cerr << "Error: could not calculate Kneser-Ney discounts for "
@@ -209,27 +209,27 @@ struct statistics {
         }
     };
 
-    statistics(uint8_t order)
+    statistics(uint64_t order)
         : m_num_ngrams(order, 0)
         , m_t(order, std::vector<uint64_t>(4, 0))
         , m_D(order, std::vector<float>(4, 0))
         , m_total_num_words(0)
         , m_unk_prob(0.0) {}
 
-    inline float D(uint8_t n, uint64_t k) const {
+    inline float D(uint64_t n, uint64_t k) const {
         assert(k > 0);
         assert(n >= 1 and n <= order());
         if (k >= 3) return m_D[n - 1].back();
         return m_D[n - 1][k - 1];
     }
 
-    inline uint64_t t(uint8_t n, uint64_t k) const {
+    inline uint64_t t(uint64_t n, uint64_t k) const {
         assert(n >= 1 and n <= order());
         assert(k > 0 and k <= 4);
         return m_t[n - 1][k - 1];
     }
 
-    inline uint64_t& num_ngrams(uint8_t n) {
+    inline uint64_t& num_ngrams(uint64_t n) {
         assert(n >= 1 and n <= order());
         return m_num_ngrams[n - 1];
     }
@@ -253,26 +253,24 @@ struct statistics {
 
     void print() {
         std::cerr << "number of ngrams:\n";
-        for (uint8_t n = 1; n <= order(); ++n) {
-            std::cerr << int(n) << "-grams: " << num_ngrams(n) << "\n";
+        for (uint64_t n = 1; n <= order(); ++n) {
+            std::cerr << n << "-grams: " << num_ngrams(n) << "\n";
         }
 
         std::cerr << "smoothing statistics:\n";
-        for (uint8_t n = 1; n <= order(); ++n) {
+        for (uint64_t n = 1; n <= order(); ++n) {
             uint64_t sum = 0;
             for (uint64_t k = 1; k <= 4; ++k) {
-                std::cerr << "t_" << int(n) << "(" << k << ") = " << t(n, k)
-                          << "\n";
+                std::cerr << "t_" << n << "(" << k << ") = " << t(n, k) << "\n";
                 sum += t(n, k);
             }
             std::cerr << "sum: " << sum << "\n" << std::endl;
         }
 
         std::cerr << "discounts:\n";
-        for (uint8_t n = 1; n <= order(); ++n) {
+        for (uint64_t n = 1; n <= order(); ++n) {
             for (uint64_t k = 1; k <= 3; ++k) {
-                std::cerr << "D_" << int(n) << "(" << k << ") = " << D(n, k)
-                          << " ";
+                std::cerr << "D_" << n << "(" << k << ") = " << D(n, k) << " ";
             }
             std::cerr << std::endl;
         }
